requirements.txt
langfuse>=3.0,<4.0
openai>=1.0
python-dotenv>=1.0
datasets>=2.0
anthropic>=0.39
fastapi>=0.115
uvicorn[standard]>=0.30
jinja2>=3.1
cachetools>=5.0
